[story-863247891] Add requireBuffer api to avoid oom in shuffle server in concurrent situation (merge request !65)

+++ common/src/main/java/com/tencent/rss/common/ShufflePartitionedBlock.java
@@ -31,7 +31,7 @@ public class ShufflePartitionedBlock {
   }
 
   public long size() {
-    return (long) length;
+    return length;
   }
 
   @Override
